third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-in-http-header-control-expected.txt

CONSOLE MESSAGE: line 4: Script executed in iframe.
CONSOLE MESSAGE: line 11: PASS: Iframe was not in a unique origin
